refactor doctor.py into a directory
