+++ lib/python/qmk/cli/doctor/__init__.py
@@ -1,0 +1,5 @@
+"""QMK Doctor
+
+Check out the user's QMK environment and make sure it's ready to compile.
+"""
+from .main import doctor
